+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `instrument_class`, `strike_price`, and `strike_price_currency` to definition
   schema
 - Added support for `imbalance` schema
+- Improved exception messages for server and client timeouts
 
 ## 0.9.0 - 2023-03-10
 - Removed `record_count` property from Bento class